refactor: address review comments
SE-6603
